renewed module structure for treebuilder
