+++ core/treebuilder/handlers.py
@@ -1,0 +1,3 @@
+class GeneralQueueHandler():
+    def __init__(self):
+        pass
