+++ core/treebuilder/wavefronts.py
@@ -1,0 +1,18 @@
+class AbstractSearchTree():
+    
+    def __init__(self):
+        """Red-Black binary search tree"""
+        pass
+
+
+class W():
+    
+    def __init__(self):
+        """Wavefront without presence of obstackles"""        pass
+
+
+class Wo():
+    
+    def __init__(self):
+        """Wavefront with presence of obstackles"""
+        pass
